+++ .gitignore
@@ -24,3 +24,6 @@
 
 .byebug_history
 .env
+
+
+dump.rdb
