+++ Gemfile
@@ -22,6 +22,8 @@ gem "octokit", "~> 4.0"
 gem 'dotenv-rails', groups: [:development, :test]
 gem "chartkick"
 gem 'jenkins_api_client'
+gem 'sidekiq'
+
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
 gem 'rails', '~> 5.1.5'
