+++ app/views/rounds/index.html.erb
@@ -4,17 +4,6 @@
 		<div class="clearfix"></div>
 		<div class="col-md-12 col-sm-12 col-xs-12">
 			<div class="x_panel">
-				<div class="x_title">
-					<h2>刷新<%= link_to ' 🔁', refresh_rounds_path, title: 'REFRESH' %></h2>
-					<ul class="nav navbar-right panel_toolbox">
-						<li>
-							<a class="collapse-link">
-								<i class="fa fa-chevron-up"></i>
-							</a>
-						</li>
-					</ul>
-					<div class="clearfix"></div>
-				</div>
 				<div class="x_content">
 					<div class="" role="tabpanel" data-example-id="togglable-tabs">
 						<ul id="myTab" class="nav nav-tabs bar_tabs" role="tablist">
