+++ spec/workers/jenkins_worker_spec.rb
@@ -1,0 +1,4 @@
+# require 'rails_helper'
+# RSpec.describe JenkinsWorker, type: :worker do
+#   pending "add some examples to (or delete) #{__FILE__}"
+# end
